pkg/destroy/gcp: break components into individual files
Previously the resources in the gcp destroy code were crowded into a
handfull of files making them difficult to manage. This change breaks
each resource type into its own file.
